Create wrapper for keysethandle and rename SelfIssuedIdToken to SelfIssuedIdTokens.
