+++ src/main/kotlin/org/wfanet/measurement/integration/common/BUILD.bazel
@@ -86,7 +86,7 @@ kt_jvm_library(
     deps = [
         ":configs",
         ":in_process_kingdom",
-        "//src/main/kotlin/org/wfanet/measurement/common/crypto/tink:self_issued_id_token",
+        "//src/main/kotlin/org/wfanet/measurement/common/crypto/tink:self_issued_id_tokens",
         "//src/main/kotlin/org/wfanet/measurement/common/identity/testing",
         "//src/main/kotlin/org/wfanet/measurement/common/identity",
         "//src/main/kotlin/org/wfanet/measurement/common/identity:principal_identity",
