+++ src/main/kotlin/org/wfanet/measurement/kingdom/deploy/gcloud/spanner/BUILD.bazel
@@ -10,7 +10,7 @@ kt_jvm_library(
         "//src/test/kotlin/org/wfanet/measurement/e2e:__subpackages__",
     ],
     deps = KINGDOM_INTERNAL_PROTOS + [
-        "//src/main/kotlin/org/wfanet/measurement/common/crypto/tink:self_issued_id_token",
+        "//src/main/kotlin/org/wfanet/measurement/common/crypto/tink:self_issued_id_tokens",
         "//src/main/kotlin/org/wfanet/measurement/kingdom/deploy/common:duchy_ids",
         "//src/main/kotlin/org/wfanet/measurement/kingdom/deploy/common/service:accounts_server_interceptor",
         "//src/main/kotlin/org/wfanet/measurement/kingdom/deploy/common/service:data_services",
