+++ src/main/kotlin/org/wfanet/measurement/kingdom/service/internal/testing/BUILD.bazel
@@ -8,7 +8,7 @@ kt_jvm_library(
     testonly = True,
     srcs = glob(["*.kt"]),
     deps = KINGDOM_INTERNAL_PROTOS + [
-        "//src/main/kotlin/org/wfanet/measurement/common/crypto/tink:self_issued_id_token",
+        "//src/main/kotlin/org/wfanet/measurement/common/crypto/tink:self_issued_id_tokens",
         "//src/main/kotlin/org/wfanet/measurement/kingdom/deploy/common/service:accounts_server_interceptor",
         "//src/main/kotlin/org/wfanet/measurement/kingdom/deploy/common:duchy_ids",
         "//src/main/kotlin/org/wfanet/measurement/kingdom/deploy/common/testing",
